refactor: reorganize src/store
