+++ src/components/Layout/index.ts
@@ -5,3 +5,4 @@ export * from './UserInfos'
 export * from './LegalNotice'
 export * from './Layout'
 export * from './LoadingServices'
+export * from './LoadingConfig'
