--- src/store/config/index.ts
@@ -1,2 +1,0 @@
-export * from './configSlice'
-export * from './LoadingConfig'
