+++ src/store/slices/config.ts
